# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant API",
        "description": "Dokumentasi API autentikasi untuk aplikasi restoran.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/orders/board": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders grouped by status (Kanban board)",
                "operationId": "b1c1ace05add50a3700ba3a7ace847fe",
                "responses": {
                    "200": {
                        "description": "Order board berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "waiting": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "processing": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "finished": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "cancelled": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data order"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}/detail": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order detail by ID",
                "operationId": "d4f742eb3640e916a5d8727969021bf5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "OR001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail order berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "order": {
                                                    "type": "object"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "payment": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order tidak ditemukan"
                    },
                    "500": {
                        "description": "Gagal mengambil detail order"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}/process-payment": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Process payment for cash orders",
                "operationId": "46c1207a35c58529d3c25d4978d8601c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "OR001"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount_paid"
                                ],
                                "properties": {
                                    "amount_paid": {
                                        "description": "Jumlah uang yang dibayarkan",
                                        "type": "number",
                                        "example": 100000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pembayaran berhasil diproses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "total": {
                                                    "type": "number"
                                                },
                                                "paid": {
                                                    "type": "number"
                                                },
                                                "change": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order tidak ditemukan"
                    },
                    "400": {
                        "description": "Jumlah pembayaran kurang"
                    },
                    "500": {
                        "description": "Gagal memproses pembayaran"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}